requirements.txt
python-pptx>=0.6.23
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=10.0.0
lxml>=4.9.0
